# hyveman-web/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "hyveman-api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hostId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ruleId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{id}": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{id}/acknowledge": {
      "post": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AlertActionRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AlertActionRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AlertActionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{id}/unacknowledge": {
      "post": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{id}/silence": {
      "post": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AlertActionRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AlertActionRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/AlertActionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{id}/unsilence": {
      "post": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rules": {
      "get": {
        "tags": [
          "Rules"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Rules"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rules/{id}": {
      "get": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "confirm",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/notification-channels": {
      "get": {
        "tags": [
          "NotificationChannels"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "NotificationChannels"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notification-channels/{id}": {
      "get": {
        "tags": [
          "NotificationChannels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "NotificationChannels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "NotificationChannels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "confirm",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/notification-channels/{id}/test": {
      "post": {
        "tags": [
          "NotificationChannels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelTestResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelTestResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelTestResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/maintenance-windows": {
      "get": {
        "tags": [
          "MaintenanceWindows"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceWindowDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceWindowDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceWindowDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MaintenanceWindows"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceWindowInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceWindowInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceWindowInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceWindowDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceWindowDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceWindowDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/retention": {
      "get": {
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionSettingsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionSettingsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionSettingsDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionSettingsInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionSettingsInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionSettingsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionSettingsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionSettingsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetentionSettingsDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertActionRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": [
              "null",
              "string"
            ]
          },
          "until": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "AlertDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ruleId": {
            "type": [
              "null",
              "string"
            ]
          },
          "ruleName": {
            "type": [
              "null",
              "string"
            ]
          },
          "hostId": {
            "type": [
              "null",
              "string"
            ]
          },
          "hostName": {
            "type": [
              "null",
              "string"
            ]
          },
          "sourceId": {
            "type": [
              "null",
              "string"
            ]
          },
          "severity": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "firstSeen": {
            "type": "string",
            "format": "date-time"
          },
          "lastSeen": {
            "type": "string",
            "format": "date-time"
          },
          "count": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "ackAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "ackReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "silenceUntil": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "resolvedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "AlertListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertDto"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}